Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/https-header-nested.html

<!DOCTYPE html>
<html>
<head>
  <script src="/js-test-resources/testharness.js"></script>
  <script src="/js-test-resources/testharnessreport.js"></script>
  <script>
    var t = async_test("Verify that nested documents receive an 'HTTPS' header.");
    window.addEventListener("message", t.step_func(function (message) {
      assert_equals(message.data.header, "1");
      t.done();
    }));
  </script>
</head>
<body>
  <iframe src="resources/post-https-header.pl"></iframe>
</body>
</html>
